+++ semcore/flags/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.2.2] - 2021-09-09
+
+### Fixed
+
+- [ts] correct types.
+
 ## [2.2.1] - 2021-8-26
 
 ### Changed